Use argument based logging configuration
- Use argument based logging configuration rather than one provided from
  the environment
- Drop "lib.rs" from previous refactor
- Remove "thiserror" and rely on "anyhow" purely
- Update dependencies
