+++ Cargo.toml
@@ -22,6 +22,7 @@ strip = true
 [dependencies]
 anyhow = { version = "1.0", features = ["backtrace"] }
 clap = { version = "4.5", features = ["derive"] }
+clap-verbosity-flag = "3.0.2"
 env_logger = { version = "0.11", features = [ "humantime" ], default-features = false }
 git2 = { version = "0.20", default-features = false }
 log = "0.4"
@@ -30,7 +31,6 @@ remain = "0.2"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 termcolor = "1.4"
-thiserror = "2.0"
 toml = "0.8"
 user_dirs = "0.2"
