+++ src/collector.rs
@@ -1,27 +1,15 @@
 //! This module contains the functionality for generating reports.
 
-use rayon::prelude::*;
 use std::collections::BTreeMap;
 use std::path::Path;
-use target::TargetCollector;
-use thiserror::Error;
 
-use crate::repository_view::{RepositoryView, RepositoryViewError, RepositoryViewResult};
+use anyhow::Result;
+use rayon::prelude::*;
+use target::TargetCollector;
+
+use crate::repository_view::RepositoryView;
 
 mod target;
-
-#[allow(missing_docs)]
-#[remain::sorted]
-#[derive(Error, Debug)]
-pub enum CollectorError {
-    #[error(transparent)]
-    FromRepositoryView(#[from] RepositoryViewError),
-    #[error(transparent)]
-    FromStdIo(#[from] std::io::Error),
-}
-
-/// The result type used when multiple kinds of errors can be encountered during collection.
-pub type CollectorResult<T> = Result<T, CollectorError>;
 
 /// This type represents a [`BTreeMap`] using an optional [`String`] for keys, which represents the
 /// parent directory for a group of reports ([`Vec<RepositoryView>`]). The values corresponding to those keys
@@ -31,7 +19,7 @@ pub type CollectorResult<T> = Result<T, CollectorError>;
 /// sorted keys.
 pub type RepositoryCollection = BTreeMap<Option<String>, Vec<RepositoryView>>;
 
-type UnprocessedRepositoryView = RepositoryViewResult<RepositoryView>;
+type UnprocessedRepositoryView = Result<RepositoryView>;
 
 /// A unit struct that provides [`Self::run()`], which is used to generated [`RepositoryCollection`].
 #[derive(Debug)]
@@ -43,7 +31,7 @@ impl RepositoryCollector {
         path: &Path,
         include_email: bool,
         include_submodules: bool,
-    ) -> CollectorResult<RepositoryCollection> {
+    ) -> Result<RepositoryCollection> {
         let unprocessed = TargetCollector::run(path.to_path_buf())?
             .par_iter()
             .map(|path| RepositoryView::new(path, include_email, include_submodules))
@@ -51,16 +39,10 @@ impl RepositoryCollector {
 
         let mut processed = RepositoryCollection::new();
         for maybe_view in unprocessed {
-            match maybe_view {
-                Ok(view) => {
-                    if let Some(mut views) =
-                        processed.insert(view.parent.clone(), vec![view.clone()])
-                    {
-                        views.push(view.clone());
-                        processed.insert(view.parent, views);
-                    }
-                }
-                Err(e) => return Err(e.into()),
+            let view = maybe_view?;
+            if let Some(mut views) = processed.insert(view.parent.clone(), vec![view.clone()]) {
+                views.push(view.clone());
+                processed.insert(view.parent, views);
             }
         }
         Ok(processed)
