+++ src/display/color.rs
@@ -1,19 +1,20 @@
 //! This module provides a harness for non-trivial displays of information to `stdout`.
 
-use gfold::Status;
 use std::io;
 use std::io::Write;
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
-use crate::config::ColorMode;
+use crate::{config::ColorMode, status::Status};
 
 /// This harness provides methods to write to `stdout`. It maps the internal [`ColorMode`] type to
 /// our dependency's [`ColorChoice`] type due to discrepancies in behavior and naming.
+#[derive(Debug)]
 pub struct ColorHarness {
     color_choice: ColorChoice,
 }
 
 impl ColorHarness {
+    /// Creates a new color harness.
     pub fn new(color_mode: ColorMode) -> Self {
         Self {
             color_choice: match &color_mode {
